move NDK into it's directory
This should help with circular dependencies
