+++ ndk/src/events/kinds/dvm/NDKDVMJobFeedback.ts
@@ -1,4 +1,4 @@
-import NDK from "../../../index.js";
+import {NDK} from "../../../ndk/index.js";
 import { NDKEvent, NostrEvent } from "../../index.js";
 import { NDKKind } from "../index.js";
