+++ ndk/src/relay/pool/index.test.ts
@@ -1,0 +1,18 @@
+import "websocket-polyfill";
+import { NDK } from "../../ndk/index.js";
+import { NDKRelay } from "../index.js";
+
+const ndk = new NDK();
+
+describe("NDKPool", () => {
+    it("refuses connecting to blacklisted relays", async () => {
+        const blacklistedRelay = new NDKRelay("wss://url1");
+        const ndk = new NDK({
+            blacklistRelayUrls: [blacklistedRelay.url],
+        });
+        const { pool } = ndk;
+        pool.addRelay(blacklistedRelay);
+
+        expect(pool.relays.size).toEqual(0);
+    });
+});
